@@ -7,9 +7,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net7.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
-	  <UseWindowsForms>true</UseWindowsForms>
     <RootNamespace>Ei_Dimension</RootNamespace>
     <PackageId>Ei_Dimension</PackageId>
     <Title>Ei_Dimension</Title>
@@ -21,6 +20,8 @@
     <Copyright>"Copyright ©  2021"</Copyright>
     <AssemblyVersion>1.6.1.1</AssemblyVersion>
     <FileVersion>1.6.1.1</FileVersion>
+    <Version>1.6.1.1</Version>
+    <AssemblyName>Ei_Dimension</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
