@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net7.0-windows</TargetFramework>
 	  <UseWPF>true</UseWPF>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
@@ -11,6 +11,7 @@
     <FileVersion>1.6.1.1</FileVersion>
     <AssemblyVersion>1.6.1.1</AssemblyVersion>
     <RootNamespace>Ei_Dimension.Language</RootNamespace>
+    <Version>1.6.1.1</Version>
   </PropertyGroup>
 
   <ItemGroup>
