@@ -5,7 +5,7 @@ namespace Ei_Dimension.Language
   public class LocExtension : Binding
   {
     public LocExtension(string name)
-        : base("[" + name + "]")
+        : base($"[{name}]")
     {
       Mode = BindingMode.OneWay;
       Source = TranslationSource.Instance;
